+++ packages/api/src/cluster/index.ts
@@ -1,14 +1,11 @@
 import { ConnectWorkerEvent, ConnectWorkerEventType } from './ConnectWorkerEvent';
-import { MemberEventType, MembershipEvent } from './MemberShip';
-import { ClusterEvent } from './ClusterEvent';
+import { ClusterEvent, MemberEventType } from './ClusterEvent';
 import { Cluster, JoinCluster, ClusterOptions } from './Cluster';
 import { MembersMap, MembersPort, MembersData } from './MemberMap';
 
 export {
   ConnectWorkerEvent,
   ConnectWorkerEventType,
-  MemberEventType,
-  MembershipEvent,
   ClusterEvent,
   Cluster,
   JoinCluster,
@@ -16,4 +13,5 @@ export {
   MembersMap,
   MembersPort,
   MembersData,
+  MemberEventType,
 };
